experiments/markets/requirements.txt
# Multi-Source Market Data Collection Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# MT5 connector
MetaTrader5>=5.0.45

# Deriv connector
websockets>=12.0
aiohttp>=3.9.0

# TradingView connector
tradingview-ta>=3.3.0

# Alpaca connector
alpaca-trade-api>=3.1.0
alpaca-py>=0.15.0

# yfinance for options data
yfinance>=0.2.30

# Data processing
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3
